requirements.txt
discord.py>=1.3.3,<1.4.0
python-dotenv>=0.12.0,<0.13.0